+++ index.html
@@ -235,5 +235,6 @@
   <script tupe="module" src="js/main.js"></script>
   <script tupe="module" src="js/utils.js"></script>
   <script tupe="module" src="js/consts.js"></script>
+  <script tupe="module" src="js/thumbnails.js"></script>
 </body>
 </html>
